+++ openpyxl2/cell/cell.py
@@ -12,6 +12,7 @@ cells using Excel's 'A1' column/row nomenclature are also provided.
 __docformat__ = "restructuredtext en"
 
 # Python stdlib imports
+from copy import copy
 import datetime
 import re
 
@@ -306,13 +307,17 @@ class Cell(StyleableObject):
         """Set value and display for hyperlinks in a cell.
         Automatically sets the `value` of the cell with link text,
         but you can modify it afterwards by setting the `value`
-        property, and the hyperlink will remain."""
-        if not isinstance(val, Hyperlink):
-            val = Hyperlink(ref="", target=val)
-        val.ref = self.coordinate
-        self._hyperlink = val
-        if self._value is None:
-            self.value = val.target or val.location
+        property, and the hyperlink will remain.
+        Hyperlink is removed if set to ``None``."""
+        if val is None:
+            self._hyperlink = None
+        else:
+            if not isinstance(val, Hyperlink):
+                val = Hyperlink(ref="", target=val)
+            val.ref = self.coordinate
+            self._hyperlink = val
+            if self._value is None:
+                self.value = val.target or val.location
 
 
     @property
@@ -384,13 +389,19 @@ class Cell(StyleableObject):
         """
         return self._comment
 
+
     @comment.setter
     def comment(self, value):
+        """
+        Assign a comment to a cell
+        """
 
         if value is not None:
-            value.parent = self
+            if value.parent:
+                value = copy(value)
+            value.bind(self)
         elif value is None and self._comment:
-            self._comment.parent = None
+            self._comment.unbind()
         self._comment = value