+++ openpyxl2/styles/stylesheet.py
@@ -87,7 +87,7 @@ class Stylesheet(Serialisable):
         self.alignments = self.cellXfs.alignments
         self.protections = self.cellXfs.prots
         self._normalise_numbers()
-        self.named_styles =  self._merge_named_styles()
+        self.named_styles = self._merge_named_styles()
 
 
     @classmethod
@@ -96,28 +96,37 @@ class Stylesheet(Serialisable):
         attrs = dict(node.attrib)
         for k in attrs:
             del node.attrib[k]
-        return  super(Stylesheet, cls).from_tree(node)
+        return super(Stylesheet, cls).from_tree(node)
 
 
     def _merge_named_styles(self):
         """
-        Merge named style names "cellStyles" with their associated styles "cellStyleXfs"
+        Merge named style names "cellStyles" with their associated styles
+        "cellStyleXfs"
         """
         named_styles = self.cellStyles.names
-        custom = self.custom_formats
-        formats = self.number_formats
+
         for style in named_styles:
-            xf = self.cellStyleXfs[style.xfId]
-            style.font = self.fonts[xf.fontId]
-            style.fill = self.fills[xf.fillId]
-            style.border = self.borders[xf.borderId]
-            if xf.numFmtId in custom:
-                style.number_format = custom[xf.numFmtId]
-            if xf.alignment:
-                style.alignment = xf.alignment
-            if xf.protection:
-                style.protection = xf.protection
+            self._expand_named_style(style)
+
         return named_styles
+
+
+    def _expand_named_style(self, named_style):
+        """
+        Bind format definitions for a named style from the associated style
+        record
+        """
+        xf = self.cellStyleXfs[named_style.xfId]
+        named_style.font = self.fonts[xf.fontId]
+        named_style.fill = self.fills[xf.fillId]
+        named_style.border = self.borders[xf.borderId]
+        if xf.numFmtId in self.custom_formats:
+            named_style.number_format = self.custom_formats[xf.numFmtId]
+        if xf.alignment:
+            named_style.alignment = xf.alignment
+        if xf.protection:
+            named_style.protection = xf.protection
 
 
     def _split_named_styles(self, wb):
@@ -165,6 +174,9 @@ def apply_stylesheet(archive, wb):
 
     wb._cell_styles = stylesheet.cell_styles
     wb._named_styles = stylesheet.named_styles
+    for ns in wb._named_styles:
+        ns.bind(wb)
+
     wb._borders = IndexedList(stylesheet.borders)
     wb._fonts = IndexedList(stylesheet.fonts)
     wb._fills = IndexedList(stylesheet.fills)