+++ openpyxl2/styles/tests/test_named_style.py
@@ -40,11 +40,11 @@ class TestNamedStyle:
 
     def test_dict(self, NamedStyle):
         style = NamedStyle()
-        assert dict(style) == {'name':'Normal', 'hidden':'0'}
+        assert dict(style) == {'name':'Normal', 'hidden':'0', 'xfId':'0'}
 
 
     def test_bind(self, NamedStyle):
-        style = NamedStyle()
+        style = NamedStyle(xfId=0)
 
         wb = Workbook()
         style.bind(wb)