+++ openpyxl2/workbook/tests/test_workbook.py
@@ -154,15 +154,14 @@ def test_get_named_range():
     wb = Workbook()
     new_sheet = wb.create_sheet()
     wb.create_named_range('test_nr', new_sheet, 'A1')
-    found_named_range = wb.get_named_range('test_nr')
-    assert found_named_range == wb.defined_names['test_nr']
+    assert wb.defined_names['test_nr'].value == 'Sheet1!A1'
 
 
 def test_remove_named_range():
     wb = Workbook()
     new_sheet = wb.create_sheet()
     wb.create_named_range('test_nr', new_sheet, 'A1')
-    wb.remove_named_range('test_nr')
+    del wb.defined_names['test_nr']
     named_ranges_list = wb.get_named_ranges()
     assert 'test_nr' not in named_ranges_list